1-print_numbers.c

+++ 0x10-variadic_functions/1-print_numbers.c
@@ -1,4 +1,4 @@
-#include <stdlib.h>
+#include <stdio.h>
 #include <stdarg.h>
 
 /**
@@ -11,10 +11,10 @@
 void print_numbers(const char *separator, const unsigned int n, ...)
 {
 	va_list args;
-	int i;
+	unsigned int i;
 
 	va_start(args, n);
-	for (i = 0; i < (int) n; i++)
+	for (i = 0; i < n; i++)
 	{
 		printf("%d", va_arg(args, int));
 		if (i < n - 1 && separator != NULL)
